[Chore #163916330] Move relations to db_models
